requirements.txt
opencv-python
streamlit
requests
pillow